chore(apps): move Error_Handler and assert_failed functions
These look the same for all apps so move them to common.
