+++ apps/servo-node/include/main.h
@@ -42,7 +42,6 @@ extern "C" {
 void HAL_TIM_MspPostInit(TIM_HandleTypeDef *htim);
 
 /* Exported functions prototypes ---------------------------------------------*/
-void Error_Handler(void);
 
 /* Private defines -----------------------------------------------------------*/
 #define PWM_PSC_1MHZ (72 - 1)
