+++ apps/servo-node/src/main.c
@@ -799,30 +799,3 @@ void StartDefaultTask(void *argument) {
     UpdatePWMDutyCycle(&pulse, &direction);
   }
 }
-
-/**
- * @brief  This function is executed in case of error occurrence.
- * @retval None
- */
-void Error_Handler(void) {
-  /* User can add his own implementation to report the HAL error return state */
-  Print("error\r\n");
-  __disable_irq();
-  while (1) {
-  }
-}
-
-#ifdef USE_FULL_ASSERT
-/**
- * @brief  Reports the name of the source file and the source line number
- *         where the assert_param error has occurred.
- * @param  file: pointer to the source file name
- * @param  line: assert_param error line source number
- * @retval None
- */
-void assert_failed(uint8_t *file, uint32_t line) {
-  /* User can add his own implementation to report the file name and line
-     number, ex: printf("Wrong parameters value: file %s on line %d\r\n", file,
-     line) */
-}
-#endif /* USE_FULL_ASSERT */
