+++ common/stm32-utils/src/utils.c
@@ -11,6 +11,13 @@
 
 extern UART_HandleTypeDef huart1;
 
+void Error_Handler(void) {
+  Print("Fatal error occured, stopping application.\r\n");
+  __disable_irq();
+  while (1) {
+  }
+}
+
 void Print(char *str) {
   HAL_UART_Transmit(&huart1, (uint8_t *)str, strlen(str), HAL_MAX_DELAY);
 }
