+++ common/stm32/src/assert.c
@@ -1,0 +1,17 @@
+#include "stm32f3xx.h"
+#include "stm32f3xx_hal_conf.h"
+
+#ifdef USE_FULL_ASSERT
+/**
+ * @brief  Reports the name of the source file and the source line number
+ *         where the assert_param error has occurred.
+ * @param  file: pointer to the source file name
+ * @param  line: assert_param error line source number
+ * @retval None
+ */
+void assert_failed(uint8_t *file, uint32_t line) {
+  /* User can add his own implementation to report the file name and line
+     number, ex: printf("Wrong parameters value: file %s on line %d\r\n", file,
+     line) */
+}
+#endif /* USE_FULL_ASSERT */
